refactor exercises
